feat: add delete confrim column  modal

+++ src/components/UI/Column/Column.tsx
@@ -15,7 +15,7 @@ interface IColumnProps {
 
 const Column: FC<IColumnProps> = ({ column }) => {
   const { title, _id: columnId, order, boardId } = column;
-  const [deleteColumn, {}] = useDeleteColumnMutation();
+  const [deleteColumn, { isLoading: isDeleting }] = useDeleteColumnMutation();
   const { data } = useGetAllTasksQuery({ boardId, columnId });
   const [createNewTask, {}] = useCreateNewTaskMutation();
   const { setLocalColumnTasks } = BoardSlice.actions;
@@ -28,10 +28,31 @@ const Column: FC<IColumnProps> = ({ column }) => {
   useEffect(() => {
     if (data) dispatch(setLocalColumnTasks([columnId, [...(data as ITask[])]]));
   }, [data]);
+  //Modal manipulations
+  ////////////////////////
+  const [isModalOpen, setModalOpen] = useState(false);
+  useEffect(() => {
+    if (!isDeleting) {
+      setModalOpen(false);
+    }
+  }, [isDeleting]);
+  ///////////////////////////////////////////////////
 
   const deleteColumnOnClick = () => {
-    deleteColumn({ boardId, columnId });
+    setModalOpen(true);
   };
+
+  function confirmDeleteColumn(e: React.MouseEvent<HTMLElement>) {
+    e.stopPropagation();
+    const target = (e.target as HTMLElement).closest('input');
+    const value = target?.value;
+    if (value == 'No' || value == 'Нет') {
+      setModalOpen(false);
+    }
+    if (value == 'Yes' || value == 'Да') {
+      deleteColumn({ boardId, columnId });
+    }
+  }
 
   const createNewTaskOnClick = () => {
     const body = {
@@ -43,6 +64,7 @@ const Column: FC<IColumnProps> = ({ column }) => {
     };
     createNewTask({ boardId, columnId, body });
   };
+
   return (
     <Draggable draggableId={columnId} index={order}>
       {(provided) => (
@@ -57,6 +79,9 @@ const Column: FC<IColumnProps> = ({ column }) => {
             deleteColumnOnClick={deleteColumnOnClick}
             createNewTaskOnClick={createNewTaskOnClick}
             dragHandleProps={provided.dragHandleProps}
+            isModalOpen={isModalOpen}
+            isDeleting={isDeleting}
+            confirmDeleteColumn={confirmDeleteColumn}
           />
           <Droppable droppableId={columnId} type="task">
             {(provided) => (
